Added required dependencies install.

+++ dev_setup.sh
@@ -1,8 +1,17 @@
 #!/usr/bin/env bash
 # Run this script to setup your development environment.
 
+if [ "$(id -u)" == "0" ]; then
+   echo "This script should not be run as root. " 1>&2
+   exit 1
+fi
+
 # Required packages.
-sudo apt-get install -y \
+read -p "Install dependencies with apt? Will need root priviliges." -n 1 -r
+echo    # move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    sudo apt-get install -y \
     python \
     python-dev \
     libxml2-dev \
@@ -10,10 +19,10 @@ sudo apt-get install -y \
     zlib1g-dev \
     python-virtualenv \
     virtualenvwrapper
-
-if [ "$(id -u)" == "0" ]; then
-   echo "This script should not be run as root. " 1>&2
-   exit 1
+    curl -sL https://deb.nodesource.com/setup_6.x | sudo -E bash -
+    sudo apt-get install nodejs
+    sudo npm install -g webpack
+    sudo npm install
 fi
 
 TOP=$(cd $(dirname $0) && pwd -L)
